Always show 9x9 grid and add more tiles

+++ src/config.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct ArenaConfig {
-    pub rows: u8,
-    pub cols: u8,
+    pub rows: u16,
+    pub cols: u16,
     pub tile_size: f32,
     pub tiles: Vec<u8>
 }
